Adding more unit tests and refactoring the driver/provider structure.
Also simplified the entityschema stuff a little to use internal __dict__ for properties.
