+++ tests/testdata/data_q_source_fail_type.yaml
@@ -1,0 +1,58 @@
+metadata:
+  version: 1
+  description: Kql Sentinel Alert Queries
+  data_environments: [LogAnalytics]
+  data_families: [SecurityAlert]
+  tags: ['alert', 'securityalert']
+defaults:
+  metadata:
+    data_source: 'security_alert'
+  parameters:
+      table:
+        description: Table name
+        type: str
+        default: 'SecurityAlert'
+      query_project:
+        description: Column project statement
+        type: str
+        default: '
+          | project-rename StartTimeUtc = StartTime, EndTimeUtc = EndTime,
+          AlertDisplayName = DisplayName, Severity = AlertSeverity
+          | extend AlertType = iif(isempty(AlertType), AlertName, AlertType)'
+      start:
+        description: Query start time
+        type: datetime
+      end:
+        description: Query end time
+        type: datetime
+      add_query_items:
+        description: Additional query clauses
+        type: str
+        default: ''
+      subscription_filter:
+        description: Optional subscription/tenant filter expression
+        type: str
+        default: 'true'
+      path_separator:
+        description: Path separator
+        type: str
+        default: '\\'
+sources:
+  missing_type:
+    description: Retrieves summary count of alerts by type
+    metadata:
+    args:
+      query: '
+        {table}
+        {query_project}
+        | where {subscription_filter}
+        | where TimeGenerated >= datetime({start})
+        | where TimeGenerated <= datetime({end})
+        | summarize alertCount=count(), firstAlert=min(TimeGenerated),
+            lastAlert=max(TimeGenerated) by AlertName
+        | order by alertCount desc
+        {add_query_items}'
+    parameters:
+      table:
+        description: Table name
+        default: 'SecurityAlert'
