+++ tests/testdata/data_q_success.yaml
@@ -1,0 +1,97 @@
+metadata:
+  version: 1
+  description: Kql Sentinel Alert Queries
+  data_environments: [LogAnalytics]
+  data_families: [SecurityAlert]
+  tags: ['alert', 'securityalert']
+defaults:
+  metadata:
+    data_source: 'security_alert'
+  parameters:
+      table:
+        description: Table name
+        type: str
+        default: 'SecurityAlert'
+      query_project:
+        description: Column project statement
+        type: str
+        default: '
+          | project-rename StartTimeUtc = StartTime, EndTimeUtc = EndTime,
+          AlertDisplayName = DisplayName, Severity = AlertSeverity
+          | extend AlertType = iif(isempty(AlertType), AlertName, AlertType)'
+      start:
+        description: Query start time
+        type: datetime
+      end:
+        description: Query end time
+        type: datetime
+      add_query_items:
+        description: Additional query clauses
+        type: str
+        default: ''
+      subscription_filter:
+        description: Optional subscription/tenant filter expression
+        type: str
+        default: 'true'
+      path_separator:
+        description: Path separator
+        type: str
+        default: '\\'
+sources:
+  query1:
+    description: Retrieves list of alerts
+    metadata:
+    args:
+      query: '
+        {table}
+        {query_project}
+        | where {subscription_filter}
+        | where TimeGenerated >= datetime({start})
+        | where TimeGenerated <= datetime({end})
+        | extend extendedProps = parse_json(ExtendedProperties)
+        | extend CompromisedEntity = tostring(extendedProps["Compromised Host"])
+        | project-away extendedProps
+        {add_query_items}'
+      uri: None
+    parameters:
+  query2:
+    description: Retrieves summary count of alerts by type
+    metadata:
+    args:
+      query: '
+        {table}
+        {query_project}
+        | where {subscription_filter}
+        | where TimeGenerated >= datetime({start})
+        | where TimeGenerated <= datetime({end})
+        | summarize alertCount=count(), firstAlert=min(TimeGenerated),
+            lastAlert=max(TimeGenerated) by AlertName
+        | order by alertCount desc
+        {add_query_items}'
+    parameters:
+  query3:
+    description: Retrieves a single alert by SystemAlertId
+    metadata:
+    args:
+      query: '
+        {table}
+        {query_project}
+        | where {subscription_filter}
+        | extend extendedProps = parse_json(ExtendedProperties)
+        | extend CompromisedEntity = tostring(extendedProps["Compromised Host"])
+        | project-away extendedProps
+        | where SystemAlertId == "{system_alert_id}"
+        {add_query_items}'
+    parameters:
+      system_alert_id:
+        description: 'The ID of the alert'
+        type: str
+      start:
+        description: Query start time
+        type: datetime
+        default: 0  # fake default to prevent this being viewed as required
+      end:
+        description: Query end time
+        type: datetime
+        default: 0  # fake default to prevent this being viewed as required
+
